+++ packages/amplify-category-api/src/provider-utils/awscloudformation/cfn-api-artifact-handler.ts
@@ -111,6 +111,8 @@ class CfnApiArtifactHandler implements ApiArtifactHandler {
       await writeResolverConfig(updates.conflictResolution, resourceDir);
     }
     const authConfig = getAppSyncAuthConfig(this.context.amplify.getProjectMeta());
+    const previousAuthConfig = _.cloneDeep(authConfig);
+
     const oldConfigHadApiKey = authConfigHasApiKey(authConfig);
     if (updates.defaultAuthType) {
       authConfig.defaultAuthentication = appSyncAuthTypeToAuthConfig(updates.defaultAuthType);
@@ -124,6 +126,7 @@ class CfnApiArtifactHandler implements ApiArtifactHandler {
         resourceDir,
         parameters: this.getCfnParameters(apiName, authConfig, resourceDir),
         authConfig,
+        previousAuthConfig,
       });
     }