+++ packages/amplify-category-auth/src/provider-utils/awscloudformation/auth-stack-builder/auth-stack-transform.ts
@@ -213,7 +213,7 @@ export class AmplifyAuthTransform extends AmplifyCategoryTransform {
       cognitoStackFileName,
     );
     // write CFN template
-    writeCFNTemplate(template, cognitostackFilePath, {
+    await writeCFNTemplate(template, cognitostackFilePath, {
       templateFormat: CFNTemplateFormat.JSON,
     });
     // write parameters.json
@@ -503,7 +503,7 @@ export class AmplifyAuthTransform extends AmplifyCategoryTransform {
       this._authTemplateObj.addCfnParameter(
         {
           type: 'String',
-          default: [],
+          default: '[]',
         },
         'hostedUIProviderCreds',
       );