+++ packages/amplify-category-auth/src/provider-utils/awscloudformation/utils/trigger-file-uploader.js
@@ -2,6 +2,7 @@ const { readdirSync, existsSync } = require('fs');
 const { createReadStream } = require('fs-extra');
 const Ora = require('ora');
 const mime = require('mime-types');
+const { stateManager } = require('amplify-cli-core');
 const sequential = require('promise-sequential');
 const { getAuthResourceName } = require('../../../utils/getAuthResourceName');
 
@@ -24,7 +25,9 @@ async function uploadFiles(context) {
     }
     const assetPath = `${authPath}/assets`;
     const env = context.amplify.getEnvInfo().envName;
-    const bucketName = `${context.amplify.readJsonFile(`${authPath}/build/parameters.json`).verificationBucketName}-${env}`;
+
+    const authParams = stateManager.getResourceParametersJson(undefined, 'auth', authResource);
+    const bucketName = `${authParams.verificationBucketName}-${env}`;
 
     if (!existsSync(assetPath)) {
       return null;