+++ packages/amplify-graphql-transformer-core/src/index.ts
@@ -1,7 +1,7 @@
 import { print } from 'graphql';
 import { EXTRA_DIRECTIVES_DOCUMENT } from './transformation/validation';
 export { GraphQLTransform, GraphQLTransformOptions, SyncUtils } from './transformation';
-export { DeploymentResources } from './transformation/types';
+export { DeploymentResources, UserDefinedSlot } from './transformation/types';
 export { validateModelSchema } from './transformation/validation';
 export {
   ConflictDetectionType,
@@ -36,4 +36,4 @@ export const getAppSyncServiceExtraDirectives = (): string => {
   return print(EXTRA_DIRECTIVES_DOCUMENT);
 };
 
-export { MappingTemplate } from './cdk-compat';
+export { MappingTemplate, TransformerNestedStack } from './cdk-compat';